+++ R/stat_pvalue_manual.R
@@ -102,6 +102,7 @@ stat_pvalue_manual <- function(
   data, label = NULL, y.position = "y.position",
   xmin = "group1", xmax = "group2", x = NULL,
   size = 3.88, label.size = size, bracket.size = 0.3,
+  bracket.nudge.y = 0,
   color = "black", linetype = 1, tip.length = 0.03,
   remove.bracket = FALSE, step.increase = 0, step.group.by = NULL,
   hide.ns = FALSE, vjust = 0, position = "identity", ...
@@ -201,11 +202,13 @@ stat_pvalue_manual <- function(
       # case when ref.group = "all"
       bracket.size = 0
     }
+
     geom_exec(
       geom_bracket, data = data, xmin = "xmin", xmax = "xmax",
       label = "label", y.position = "y.position", vjust = "vjust",
       group = 1:nrow(data),  tip.length =  tip.length,
-      label.size = label.size, size = bracket.size, color = color,
+      label.size = label.size, size = bracket.size,
+      bracket.nudge.y = bracket.nudge.y, color = color,
       linetype = linetype, step.increase = step.increase,
       step.group.by = step.group.by, position = position, ...
     )